+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fixed parsing username+password git remote URLs in `launch.beaker` module.
 - Cluster names in Beaker have changed.
+- Fixed mixture rounding error with `SourceMixtureDataset`, which was previously causing samples to be repeated at the end of training.
 
 ### Changed
 
@@ -94,6 +95,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Optimization: avoid redundant calls to `model.train()` in `TransformerTrainModule`.
 - `NumpyDatasetConfig.expand_glob` now works with remote directories.
 - Fixed Attention block sharding when TP and head-wise QK norm are both applied.
+
 
 ## [v2.1.0](https://github.com/allenai/OLMo-core/releases/tag/v2.1.0) - 2025-04-14